Reserve a dedicated arena for every construction of mmap_address_map in order to reduce memory usage for HEAP_PROFILE_MMAP.
Large memory usage is observed for heap profiling while profiling mmap().  This change fixes the bloat by reserving a dedicated arena for every construction of mmap_address_map.

This change is planned to be upstreamed.  If another approach is chosen in the upstream, this change will be replaced.

BUG=123759
TEST=run all existing tests.

Review URL: http://codereview.chromium.org/9963095

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@132771 0039d316-1c4b-4281-b951-d872f2087c98