issue-368: unmap free spans and retry before growing heap
Because unmapped spans are not coalesced with normal spans it's
possible that we indeed have a large enough free span, but we fail to
see that because we always consider unmapped and normal spans
separately. That behavior is more likely for larger spans.

Given that heap grows is not expected to be frequent event I believe
it's more than ok to spend some time trying to unmap everything we
have.

For same reason I believe even worst case is fine too. Basically new
code will cause whatever spans we had free but not big enough to
satisfy current allocation to incur minor page faults next time they
are allocated. But I believe alternative of memory blowup (where minor
page faults will occur anyway for newly grown span) is worse.