synchronized Sampler implementation with Google-internal version
This is mostly dropping FastLog2 which was never necessary for
performance, and making sampler to be called always, even if sampling
is disabled (this benefits more for always-sampling case of Google
fork).

We're also getting TryRecordAllocationFast which is not used yet, but
will be as part of subsequent fast-path speedup commit.