cleaned up and fixed profiler_unittest
This includes dropping unnecessary "multiplicity" (i.e. we had
profiler{1,2,3,4}_unittest binaries). Apparently attempt was made to
test with/without threads and with different linking options. But
threads are "always on" nowadays and those were not really different
linking options (at least with current-day libtool).

Another issue was that modern pprof outputs units of time (so 530ms
instead of 530) which confused expr thingy that we use in our shell
script. This is now fixed (by cutting ms suffix via sed).